fix(paddleocr): address review comments and fix critical bugs
Address review comments from PR #3247:
- Rename get_sdk_client to get_api_client_config to better reflect
  that it returns HTTP API config, not an SDK client
- Update base_url help text: replace 'self-hosted deployments' with
  'special gateway scenarios' since the full async service is not
  open-sourced
- Make base_url a keyword-only argument for clarity
- Move build_*_options functions into their respective tool files

Bug fixes:
- Fix _poll_job resultUrl parsing: API returns result URL under
  data.resultUrl.jsonUrl (not data.resultJsonUrl). Added fallback
  chain to handle both formats.
- Filter file/fileType keys from options sent to API
