+++ tools/paddleocr/manifest.yaml
@@ -1,4 +1,4 @@
-version: 0.2.7
+version: 0.2.8
 type: plugin
 author: langgenius
 name: paddleocr
