+++ tools/paddleocr/tools/text_recognition.py
@@ -5,12 +5,24 @@ from dify_plugin import Tool
 from dify_plugin.entities.tool import ToolInvokeMessage
 
 from tools.utils import (
-    build_ocr_options,
     call_paddleocr_api,
+    camel_to_snake,
     cleanup_temp_file,
-    get_sdk_client,
+    get_api_client_config,
     normalize_file_input,
 )
+
+_SKIP_KEYS = {"file", "fileType"}
+
+
+def build_ocr_options(params: dict[str, Any]) -> dict[str, Any]:
+    """Build OCR options dict from parameters using dynamic conversion."""
+    options_dict = {}
+    for api_name, value in params.items():
+        if value is None or api_name in _SKIP_KEYS:
+            continue
+        options_dict[camel_to_snake(api_name)] = value
+    return options_dict
 
 
 class TextRecognitionTool(Tool):
@@ -22,7 +34,7 @@ class TextRecognitionTool(Tool):
             )
         access_token = self.runtime.credentials["aistudio_access_token"]
 
-        # Get base_url (optional, uses SDK default if not provided)
+        # Get base_url (optional, uses default if not provided)
         base_url = self.runtime.credentials.get("base_url")
 
         # Normalize file input - returns (input_value, is_temp_file, file_type_code)
@@ -35,7 +47,7 @@ class TextRecognitionTool(Tool):
             options = build_ocr_options(tool_parameters)
 
             # Get API client config
-            client_config = get_sdk_client(access_token, base_url)
+            client_config = get_api_client_config(access_token, base_url=base_url)
 
             # Call API
             if file_input.startswith(("http://", "https://")):
